Create utils directory and utils/math.hxx
