+++ src/utils/math.hxx
@@ -1,37 +1,10 @@
-#ifndef UTILS_HXX_
-#define UTILS_HXX_
-
-#include <GL/gl.h>
-
-#include <SFML/Graphics.hpp>
+#ifndef MATH_HXX_
+#define MATH_HXX_
 
 #include <vector>
-#include <string>
 
-/* Function used to load files like shader source code
-  \param path The path to the file that you want to load
-  \return the string containing the content of the file
-*/
-std::string loadFile(std::string path);
-
-template<typename Out>
-void split(const std::string &s, char delim, Out result);
-
-/* Function used to split string into a list of strings using a delimiter
-  \param s The string that you want to split
-  \param delim The delimiter used to split the string
-  \return a vector of substrings
-*/
-std::vector<std::string> split(const std::string &s, char delim);
-
-/* Function used to display OpenGL errors
-  \param file The name of current file
-  \param line The line where displayGLErrors is called
-  \return a boolean which is true if there is errors, false otherwise
-*/
-bool _displayGLErrors(const char *file, int line);
-
-#define displayGLErrors() _displayGLErrors(__FILE__, __LINE__)
+#include <GL/gl.h>
+#include <SFML/Graphics.hpp>
 
 /* Generate and return a perspective matrix. Inspired from the gluPerspective
   function, but it only creates the matrix and returns it, it doesn't call
